LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/memory_redundancy64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory_redundancy64.wast:5"
FAIL #3 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) true but got (string) "wrapper"
PASS #4 Reinitialize the default imports
FAIL #5 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 52: Memory64 is not enabled at {loc} expected true got false
FAIL #6 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #7 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #8 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:13:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #9 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #10 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:19:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #11 Test that WebAssembly compilation succeeds assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:22:31
run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:224:37
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:22:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got "wrapper"
FAIL #12 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:22:24
run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:224:37
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:22:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #13 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:22:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #14 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:25:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory_redundancy64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory_redundancy64.wast.js:30:3 expected true got false
